wifi: ieee80211: split HT definitions out
[ Upstream commit fdc1c141f3ef4dc94e3880e973061681843f62c0 ]

The ieee80211.h file has gotten very long, continue splitting
it by putting HT definitions into a separate file.

Link: https://patch.msgid.link/20251105153843.7532471178d0.Id956a5433ad8658e4e5c0272dbcbb59587206142@changeid
Stable-dep-of: cb0caadb64ca ("wifi: ieee80211: fix definition of EHT-MCS 15 in MRU")
(cherry picked from commit 4d5caab09dabaf293b2fd8f9202ff4a2f3111b65)
